Move exec.go from dockertools to dockershim
