+++ pkg/kubelet/dockertools/docker_manager.go
@@ -287,26 +287,6 @@ func GetUserFromImageUser(id string) string {
 	return id
 }
 
-type dockerExitError struct {
-	Inspect *dockertypes.ContainerExecInspect
-}
-
-func (d *dockerExitError) String() string {
-	return d.Error()
-}
-
-func (d *dockerExitError) Error() string {
-	return fmt.Sprintf("Error executing in Docker Container: %d", d.Inspect.ExitCode)
-}
-
-func (d *dockerExitError) Exited() bool {
-	return !d.Inspect.Running
-}
-
-func (d *dockerExitError) ExitStatus() int {
-	return d.Inspect.ExitCode
-}
-
 // RewriteResolvFile rewrites resolv.conf file generated by docker.
 // Exported for reusing in dockershim.
 func RewriteResolvFile(resolvFilePath string, dns []string, dnsSearch []string, useClusterFirstPolicy bool) error {
